provider details with router
