+++ client/package.json
@@ -8,6 +8,7 @@
     "@fortawesome/free-regular-svg-icons": "^5.9.0",
     "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/react-fontawesome": "^0.1.4",
+    "google-maps-react": "^2.0.2",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
     "react-datepicker": "^2.8.0",
