+++ client/src/components/App.jsx
@@ -9,17 +9,15 @@ import Footer from './common/Footer';
 import Home from './home/Home';
 
 import ActivityMainPage from './activity/ActivityMainPage';
-import Activity from './activity/Activity';
+import ActivityDetails from './activity/ActivityDetails';
 import AddActivity from './activity/AddActivity';
-import Providers from './provider/Providers';
+import ProviderMainPage from './provider/ProviderMainPage';
 import ProviderDetails from './provider/ProviderDetails';
 import AddProvider from './provider/AddProvider';
 import EventMainPage from './event/EventMainPage';
 import ThingsToDoMainPage from './things-to-do/ThingsToDoMainPage';
 import Users from './user/Users';
 import Tutorials from './tutorial/Tutorials';
-import GreetingClass from './tutorial/GreetingClass';
-import GreetingHooks from './tutorial/GreetingHooks';
 
 import NotFound from './shared/NotFound'
 
@@ -36,17 +34,15 @@ const App = () => (
                         <Redirect from="/home" to="/" />
                         <Route exact path="/" component={Home} />
                         <Route exact path="/activities" component={ActivityMainPage} />
-                        <Route path="/activity/:id" component={Activity} />
+                        <Route path="/activity/:id" component={ActivityDetails} />
                         <Route path="/activities/new" component={AddActivity} />
-                        <Route exact path="/providers" component={Providers} />
+                        <Route exact path="/providers" component={ProviderMainPage} />
                         <Route path="/provider/:id" component={ProviderDetails} />
                         <Route path="/providers/new" component={AddProvider} />
                         <Route exact path="/events" component={EventMainPage} />
                         <Route exact path="/things-to-do" component={ThingsToDoMainPage} />
                         <Route path="/users" component={Users} />
                         <Route exact path="/tutorials" component={Tutorials} />
-                        <Route path="/tutorials/greeting_class" component={GreetingClass} />
-                        <Route path="/tutorials/greeting_hooks" component={GreetingHooks} />
                         <Route component={NotFound} />
                     </Switch>
                     <Footer />
