+++ client/src/components/common/Header.jsx
@@ -24,10 +24,11 @@ export default function Header() {
                                 <div className="navbar-nav">
                                     <HeaderLink to="/">Home</HeaderLink>
                                     <HeaderLink to="/activities">Camps & Classes</HeaderLink>
-                                    <HeaderLink to="/providers">[Provider]</HeaderLink>
+                                    <HeaderLink to="/providers">Providers</HeaderLink>
                                     <HeaderLink to="/things-to-do">Things To Do</HeaderLink>
                                     <HeaderLink to="/events">Events</HeaderLink>
                                     <HeaderLink to="/blogs">Blog</HeaderLink>
+                                    <HeaderLink to="/about">About</HeaderLink>
                                     <HeaderLink to="/users">[User]</HeaderLink>
                                     <HeaderLink to="/tutorials">Sandbox</HeaderLink>
                                 </div>
