--- client/src/components/provider/Providers.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import ProviderList from './ProviderList';
-
-export default function Providers() {
-    return (
-        <ProviderList />
-    );
-}
